WIP but as there is no support for non-terminal variadic arguments (so source_location can be a default argument) it is still not possible...

+++ Render/Render/Core/ISwapChain.cpp
@@ -28,7 +28,7 @@ uint32_t ISwapChain::GetSignalValue()
 
 	frameCounters[bufferIndex] = currentFrameIndex;
 
-	return frameCounters[bufferIndex];
+	return (uint32_t)frameCounters[bufferIndex];
 }
 }
 }
